+++ Pipeline_2/utils.py
@@ -1,0 +1,13 @@
+
+from itertools import chain, repeat
+
+def compute_pows():
+    print("Not today")
+    
+    
+def get_user_input(valid_response, prompt, err_prompt):
+    prompts = chain([prompt], repeat(err_prompt))
+    replies = map(input, prompts)
+    lowercased_replies = map(str.lower, replies)
+    stripped_replies = map(str.strip, lowercased_replies)
+    return next(filter(valid_response.__contains__, stripped_replies))
